Create README.md
# WEB Part C, Rate Your Lecture - RYL
**Group 10 - Amir Olizkey, Shay Twieg**

## Project Overview
This web application, "Rate Your Lecturer - RYL," is designed for college and university students to provide a comprehensive platform for rating and giving feedback on lecturers. The app aims to fulfill students' desire for objective information on lecturers and courses, allowing them to easily share their experiences, give ratings, and leave comments. Lecturers can view their ratings and access statistical information on feedback, fostering a supportive and knowledge-sharing community. Each lecturer will have a profile page with personal details, including name, photo (if approved), teaching institutions, subjects taught, degrees, years of experience, interests, overall rating, and student comments. Additionally, each course will have its own page with a brief description and a display of all lecturers teaching the course by institution.

## User Guide

### Visiting the Home Page
1. Navigate to the home page.
2. From the home page you can enter the pages of relevant lecturers or courses and look at the websites of recognized universities.

### Register
3. Click on the 'הירשם' button to register as a new user.
4. Fill out the registration form with your personal details.
5. Submit the form to create your new account.

### Logging In
6. If you are an existing user, click the 'התחבר' button.
7. Enter your login credentials (Email and password).
8. Click on 'כניסה' to access your account.

### Searching for Lecture or Course
9. Once logged in, navigate to the home page via the navigation bar.
10. Use the search button to find lectures or courses that fit your list.
11. Examples of lecturers in DB (you can search it): סנייפ, מיכל כהן, שרית אזולאי, חיים לוי
12. Examples of courses in DB (you can search it): ריפוי בכשפים ושיקויים רפואיים, פילוסופיה של המוסר, אבטחת מידע וסייבר, מבוא לבינה מלאכותית

### Viewing and Rating Lectures
13. You can rate the lecture based on your experience.

### Participation in Discussions 
14. Navigate to the 'פורום דיונים' page by the navigator.
15. You can view all discussions and write a new discussion in the appropriate place.
16. If you want to write a comment to an existing discussion, click on the discussion and you can view the comments and write a new comment on this discussion.

### Viewing Events at Various Universities 
17. Navigate to the 'אירועים וחדשות' page by the navigator.
18. You will be able to watch the events that are held at different universities, see what dates they are and enter the link to the event website.

### Updating Personal Details
19. To update your personal details, navigate to 'פרופיל משתמש' via the navigation bar.
20. You can edit your details here.

### Logging Out
21. To log out of your account, click the 'Logout' button found in the navigation bar.
22. You will be redirected back to the home page.

## Screenshots

### Home Page
The home page of the site where you can login, register, and view lecturers or courses pages by searching for them in the appropriate place.
![Home Page (Visitor)](./static/media/screenshots/homePageVisiter.png "Home Page (Visitor)")
![Home Page (User)](./static/media/screenshots/homePageUser.png "Home Page (User)")

### Register Page
For new users to sign up.
![Register Page](./static/media/screenshots/register.png "Register Page")

### Login Page
For existing users to sign in.
![Login Page](./static/media/screenshots/login.png "Login Page")

### Course Page
You can view the details of the course, its description, academic field, and lecturers who teach it at which university.
![Course Page](./static/media/screenshots/course.png "Course Page")

### Lecturer Page
You can view the lecturer's details, his rating, and write a new comment if you are a registered user.
![Lecturer Page](./static/media/screenshots/lecturer.png "Lecturer Page")

### Forum Page
You can view all the discussions and also add a new discussion if you are a registered user.
![Forum Page](./static/media/screenshots/forum.png "Forum Page")

### Discussion Page
You will be able to enter the discussion, see the comments on it, and comment a new comment if you are a registered user.
![Discussion Page](./static/media/screenshots/discussion.jpg "Discussion Page")

### Events and News Page
You will be able to watch a variety of events held at the various universities.
![Events Page](./static/media/screenshots/events.png "Events Page")

### Profile Page
Here the user will be able to see his details and update them when necessary.
![Profile Page](./static/media/screenshots/profile.png "Profile Page")

### Logout Page
Allows users to disconnect and return to the home page.
![Logout Page](./static/media/screenshots/logout.png "Logout Page")